Backend/server/requirements.txt
fastapi
uvicorn
pydantic
pandas
python-dotenv
litellm
google-adk
google-auth
